requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
langchain>=0.1.0
pillow>=10.0.0
python-dotenv>=1.0.0